Python: reorganized and added to tests
